Ajusting dialog interface

+++ client/pages/transportation.tsx
@@ -2,6 +2,7 @@ import Head from "next/head";
 
 import { useState, useEffect, FormEvent } from "react";
 import * as Dialog from "@radix-ui/react-dialog";
+import * as AlertDialog from "@radix-ui/react-alert-dialog";
 
 import { Nav } from "./components/Nav";
 import { Input } from "./components/Input";
@@ -12,6 +13,7 @@ import {
   PlusIcon,
 } from "@heroicons/react/24/outline";
 import axios from "axios";
+import { PlusSmallIcon } from "@heroicons/react/24/solid";
 
 type DataProps = {
   id: string;
@@ -30,6 +32,7 @@ export default function Transportation() {
   const [items, setItems] = useState<DataProps[]>([]);
   const [open, setIsOpen] = useState(false);
   const [query, setQuery] = useState("");
+  const [deleteDialogOpen, setDeleteDialogOpen] = useState(false);
 
   useEffect(() => {
     axios("http://localhost:4000/deliveryitems").then((response) =>
@@ -42,8 +45,8 @@ export default function Transportation() {
       const response = axios.delete(
         `http://localhost:4000/deliveryitems/${id}`
       );
+      setDeleteDialogOpen(false);
       setItems(items.filter((p) => p.id !== id));
-      alert("Item deleted with success");
     } catch (error) {
       console.log(error);
       alert("Error while trying to delete the item!");
@@ -80,20 +83,21 @@ export default function Transportation() {
     return open ? (
       <Dialog.Portal>
         <Dialog.Overlay className="fixed inset-0 bg-black/60" />
-        <Dialog.Content className="fixed top-1/2 left-1/2 w-[480px] -translate-x-1/2 -translate-y-1/2 rounded-lg border border-zinc-700 bg-[#0d1117] py-8 px-10 text-white shadow-lg shadow-black/25">
-          <Dialog.Title className="text-center">
-            Add a new delivery
+        <Dialog.Content className="fixed top-1/2 left-1/2 w-[450px] -translate-x-1/2 -translate-y-1/2 rounded-lg border border-zinc-700 bg-[#0d1117] py-8 px-10 text-white shadow-lg shadow-black/25">
+          <Dialog.Title className="flex items-center gap-1 text-xl">
+            New delivery
+            <PlusSmallIcon className="h-7 w-7" />
           </Dialog.Title>
-          <Dialog.Description className="pt-0.5 text-center text-sm text-[#8B949E]">
-            Fill in all fields
+          <Dialog.Description className="pt-0.5 text-sm text-[#8B949E]">
+            Fill in all fields to create a new delivery in our database.
           </Dialog.Description>
-          <form onSubmit={handleCreateItem} className="mt-8">
+          <form onSubmit={handleCreateItem} className="mt-6">
             <div className="flex flex-col gap-4">
-              <div className="flex justify-evenly">
+              <div className="flex justify-between">
                 <div>
                   <label
                     htmlFor="carNumber"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Del. Number
                   </label>
@@ -108,7 +112,7 @@ export default function Transportation() {
                 <div>
                   <label
                     htmlFor="dateTime"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Date / Time
                   </label>
@@ -122,11 +126,11 @@ export default function Transportation() {
                 </div>
               </div>
 
-              <div className="flex justify-evenly">
+              <div className="flex justify-between">
                 <div>
                   <label
                     htmlFor="tu"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     TU
                   </label>
@@ -141,7 +145,7 @@ export default function Transportation() {
                 <div>
                   <label
                     htmlFor="carrier"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Carrier
                   </label>
@@ -154,11 +158,11 @@ export default function Transportation() {
                   />
                 </div>
               </div>
-              <div className="flex justify-evenly">
+              <div className="flex justify-between">
                 <div>
                   <label
                     htmlFor="shipTo"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Ship to
                   </label>
@@ -173,7 +177,7 @@ export default function Transportation() {
                 <div>
                   <label
                     htmlFor="weight"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Weight
                   </label>
@@ -186,11 +190,11 @@ export default function Transportation() {
                   />
                 </div>
               </div>
-              <div className="flex justify-evenly">
+              <div className="flex justify-between">
                 <div>
                   <label
                     htmlFor="numberOfItems"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Number of items
                   </label>
@@ -205,7 +209,7 @@ export default function Transportation() {
                 <div>
                   <label
                     htmlFor="pickingStatus"
-                    className="text-xs font-semibold text-white"
+                    className="text-sm font-semibold text-white"
                   >
                     Status
                   </label>
@@ -237,6 +241,29 @@ export default function Transportation() {
           </form>
         </Dialog.Content>
       </Dialog.Portal>
+    ) : null;
+  }
+
+  function deleteDialog() {
+    return deleteDialogOpen ? (
+      <AlertDialog.Portal>
+        <AlertDialog.Overlay className="white fixed inset-0 bg-black bg-opacity-5" />
+        <AlertDialog.Content className="fixed top-1/2 left-1/2 w-[480px] -translate-x-1/2 -translate-y-1/2 rounded-lg border border-zinc-700 bg-[#0d1117] py-8 px-10 text-white shadow-lg shadow-black/25">
+          <AlertDialog.Title className="text-xl">
+            Delivery successfully deleted
+          </AlertDialog.Title>
+          <AlertDialog.Description className="pt-0.5  text-sm text-[#8B949E]">
+            This delivery has been successfully deleted.
+          </AlertDialog.Description>
+          <div>
+            <AlertDialog.Action asChild>
+              <button className="mt-4 flex h-[40px] w-[120px] items-center justify-center gap-1 rounded-lg border border-zinc-700 bg-[#0d1117] pt-0 text-sm font-bold text-white duration-300  hover:border-[#1f61fb] hover:bg-[#131922] hover:text-white  hover:outline-none hover:ring-1 hover:ring-[#1f61fb] hover:transition-colors">
+                Got it, thanks!
+              </button>
+            </AlertDialog.Action>
+          </div>
+        </AlertDialog.Content>
+      </AlertDialog.Portal>
     ) : null;
   }
 
@@ -367,14 +394,22 @@ export default function Transportation() {
                         <td className="py-4 px-6">{items.weight}</td>
                         <td className="py-4 px-6">{items.numberOfItems}</td>
                         <td className="py-4 px-6">{items.pickingStatus}</td>
-                        <td className="py-4 px-6 text-right">
-                          <button
-                            onClick={() => handleDelete(items.id)}
-                            className="font-medium text-white hover:underline"
-                          >
-                            Delete
-                          </button>
-                        </td>
+                        <AlertDialog.Root
+                          open={deleteDialogOpen}
+                          onOpenChange={setDeleteDialogOpen}
+                        >
+                          <td className="py-4 px-6 text-right">
+                            {deleteDialog()}
+                            <AlertDialog.Trigger>
+                              <button
+                                onClick={() => handleDelete(items.id)}
+                                className="font-medium text-white hover:underline"
+                              >
+                                Delete
+                              </button>
+                            </AlertDialog.Trigger>
+                          </td>
+                        </AlertDialog.Root>
                       </tr>
                     );
                   })}
